Move documentSelector from state to config
