+++ src/config.ts
@@ -4,9 +4,19 @@ const KEYBINDINGS_ENABLED_CONTEXT_KEY = 'calva:keybindingsEnabled';
 
 type ReplSessionType = 'clj' | 'cljs';
 
+// include the 'file' and 'untitled' to the
+// document selector. All other schemes are
+// not known and therefore not supported.
+const documentSelector = [
+    { scheme: 'file', language: 'clojure' },
+    { scheme: 'jar', language: 'clojure' },
+    { scheme: 'untitled', language: 'clojure' }
+];
+
 export {
     REPL_FILE_EXT,
     KEYBINDINGS_ENABLED_CONFIG_KEY,
     KEYBINDINGS_ENABLED_CONTEXT_KEY,
+    documentSelector,
     ReplSessionType
 }
